requirements.txt
ffmpeg-python
PyQt5